% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagotchi_barplot.R
\name{datagotchi_barplot}
\alias{datagotchi_barplot}
\title{Create a standardized Datagotchi bar plot}
\usage{
datagotchi_barplot(
  data,
  x,
  y,
  fill,
  y_title = "Proportion",
  percents = TRUE,
  x_rename = NULL,
  fill_rename = NULL,
  logo = TRUE,
  colorset = "issues",
  n = TRUE,
  caption = NULL,
  theme = "datagotchi_light",
  save_path = NULL,
  weight = NULL
)
}
\arguments{
\item{data}{A data frame containing the data to plot}

\item{x}{Character. The x-axis variable name}

\item{y}{Character. The y-axis variable name (if computing proportions, this will be the variable to compute proportions from)}

\item{fill}{Character. The fill variable name}

\item{y_title}{Character. Title for y-axis}

\item{percents}{Logical. Whether to display values as percentages}

\item{x_rename}{Named vector. Renaming values for x variable}

\item{fill_rename}{Named vector. Renaming values for fill variable}

\item{logo}{Logical. Whether to add the Datagotchi logo}

\item{colorset}{Character. Name of the color set to use ("issues", "parties", "custom")}

\item{n}{Logical. Whether to display sample size in caption}

\item{caption}{Character. Caption text (will be appended with n if n=TRUE)}

\item{theme}{Character. Name of the theme to use}

\item{save_path}{Character. Path where to save the plot}

\item{weight}{Character. Optional. Name of the weight variable for weighted calculations}
}
\description{
Create a standardized Datagotchi bar plot
}
