% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_graph.R
\name{make_graph}
\alias{make_graph}
\title{Create an interactive graph making experience}
\usage{
make_graph(data)
}
\arguments{
\item{data}{A data frame containing the data to plot}
}
\description{
Create an interactive graph making experience
}
